game-flow: combine item drop support
This adds support for game-flow defined data drops in TR2, similar
to the OG levels in TR1. The setting will default to false to allow
regular item pickups to work as normal without having to reconfigure
anything.
