+++ data/tr2/ship/cfg/tr2-gm/gameflow.json5
@@ -7,6 +7,8 @@
     "savegame_fmt_bson": "save_trgm_%02d.dat",
 
     "demo_version": false,
+    "enable_tr2_item_drops": true,
+    "convert_dropped_guns": true,
 
     "title": {
         "path": "data/title_gm.tr2",
